Oiram fully playable on calculator! Still has a little cleaning up to do Fixed multi file find for device (BFile_Find only works ONCE on Prizm for some reason) Changed argument on SetPalette offset to allow you to effect the whole palette

+++ src/loadscreen.h
@@ -23,7 +23,7 @@ void load_progress(void);
 void set_load_screen(void);
 void set_level(char *name, uint8_t level);
 
-extern unsigned int *warp_info;
+extern uint8_t *warp_info;
 extern unsigned int warp_num;
 
 #endif
